fixed namespaces (mostly for the Common directories)
