+++ backend/backend.tests/Common/DependOnOtherComponentOnlyThroughApiNamespaceCustomCondition.cs
@@ -74,13 +74,11 @@ namespace backend.tests.Common
                         string targetFullNamespace = targetType.Namespace.FullName;
                         Regex apiRegex = new Regex(@"\.Api(?:$|\.)");
                         bool isApiDependency = apiRegex.IsMatch(targetFullNamespace);
-                        Regex commonRegex = new Regex(@"\.Common(?:$|\.)");
-                        bool isCommonDependency = commonRegex.IsMatch(targetFullNamespace);
 
-                        if (!isApiDependency && !isCommonDependency)
+                        if (!isApiDependency)
                         {
                             overallPassForSourceType = false;
-                            failureDetails.Add($"{failedTestCount + 1} - depends on {targetType.FullName} (in component {targetComponentName}) which is not in an 'Api' or 'Common' namespace of that component");
+                            failureDetails.Add($"{failedTestCount + 1} - depends on {targetType.FullName} (in component {targetComponentName}) which is not in an 'Api' namespace of that component");
 
                             failedTestCount++;
                         }
