+++ backend/backend/backend.csproj
@@ -19,4 +19,10 @@
         </Content>
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="src\AccountComponent\Common\Api\" />
+      <Folder Include="src\SignalingComponent\Common\Api\" />
+      <Folder Include="src\WebSocketComponent\Common\Api\" />
+    </ItemGroup>
+
 </Project>
