+++ backend/backend/src/AccountComponent/Common/Api/DTOs/AccountCreateDTO.cs
@@ -1,19 +1,17 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 
-namespace backend.AccountComponent.Common.DTOs;
+namespace backend.AccountComponent.Common.Api.DTOs;
 
-public class AccountRetrieveDto
+public class AccountCreateDto
 {
     [Required]
+    [StringLength(100, MinimumLength = 3)]
     [JsonPropertyName("username")]
     public required string DisplayName { get; set; }
     
     [Required]
+    [StringLength(100, MinimumLength = 6)]
     [JsonPropertyName("password")]
     public required string Password { get; set; }
-    
-    [Required]
-    [JsonPropertyName("id")]
-    public required int Id { get; set; }
 }
