+++ backend/backend/src/AccountComponent/Common/Api/DTOs/LoginResponse.cs
@@ -1,0 +1,3 @@
+﻿namespace backend.AccountComponent.Common.Api.DTOs;
+
+public record LoginResponse(string Message);
