+++ backend/backend/src/AccountComponent/Common/Api/DTOs/StatusResponse.cs
@@ -1,0 +1,3 @@
+namespace backend.AccountComponent.Common.Api.DTOs;
+
+public record StatusResponse(bool Status);
