--- backend/backend/src/AccountComponent/Common/DTOs/LoginResponse.cs
@@ -1,3 +1,0 @@
-﻿namespace backend.AccountComponent.Common.DTOs;
-
-public record LoginResponse(string Message);
