--- backend/backend/src/AccountComponent/Common/DTOs/StatusResponse.cs
@@ -1,3 +1,0 @@
-namespace backend.AccountComponent.Common.DTOs;
-
-public record StatusResponse(bool Status);
