+++ backend/backend/src/AccountComponent/Dataaccess/Impl/AccountRepository.cs
@@ -1,4 +1,4 @@
-﻿using backend.AccountComponent.Common.DTOs;
+﻿using backend.AccountComponent.Common.Api.DTOs;
 using backend.AccountComponent.Dataaccess.Api.Entity;
 using backend.AccountComponent.Dataaccess.Api.Repo;
 using Npgsql;
