+++ backend/backend/src/AccountComponent/Logic/Impl/AccountCreationHandler.cs
@@ -1,6 +1,6 @@
 using System.Text.Json;
-using backend.AccountComponent.Common.DTOs;
-using backend.AccountComponent.Common.Exception;
+using backend.AccountComponent.Common.Api.DTOs;
+using backend.AccountComponent.Common.Api.Exception;
 using backend.AccountComponent.Dataaccess.Api.Entity;
 using backend.AccountComponent.Dataaccess.Api.Repo;
 using backend.AccountComponent.Logic.Api;
