+++ backend/backend/src/SignalingComponent/Common/Api/DTOs/EstablishConnectionMessage.cs
@@ -1,7 +1,7 @@
 using System.Text.Json.Serialization;
-using backend.WebSocketComponent.Common.DTOs.Api;
+using backend.WebSocketComponent.Common.Api.DTOs;
 
-namespace backend.SignalingComponent.Common.DTOs;
+namespace backend.SignalingComponent.Common.Api.DTOs;
 
 public class EstablishConnectionMessage : ITypedMessage
 {
