+++ backend/backend/src/WebSocketComponent/Common/Api/DTOs/ClientTokenMessage.cs
@@ -1,7 +1,6 @@
 using System.Text.Json.Serialization;
-using backend.WebSocketComponent.Common.DTOs.Api;
 
-namespace backend.WebSocketComponent.Common.DTOs;
+namespace backend.WebSocketComponent.Common.Api.DTOs;
 
 public class ClientTokenMessage : ITypedMessage
 {
