+++ backend/backend/src/WebSocketComponent/Common/Api/DTOs/MessageHandlerDelegate.cs
@@ -1,5 +1,5 @@
 using System.Text.Json;
 
-namespace backend.WebSocketComponent.Common.DTOs;
+namespace backend.WebSocketComponent.Common.Api.DTOs;
 
 public delegate Task MessageHandlerDelegate(string clientToken, JsonElement messageData);
