+++ backend/backend/src/WebSocketComponent/Common/Api/DTOs/TypedMessageHandlerDelegate.cs
@@ -1,3 +1,3 @@
-namespace backend.WebSocketComponent.Common.DTOs.Api;
+namespace backend.WebSocketComponent.Common.Api.DTOs;
 
 public delegate Task TypedMessageHandlerDelegate<T>(string clientToken, T messageData);
