+++ backend/backend/src/WebSocketComponent/Logic/Impl/WebSocketHandler.cs
@@ -3,8 +3,7 @@ using System.Diagnostics;
 using System.Net.WebSockets;
 using System.Text;
 using System.Text.Json;
-using backend.WebSocketComponent.Common.DTOs;
-using backend.WebSocketComponent.Common.DTOs.Api;
+using backend.WebSocketComponent.Common.Api.DTOs;
 using backend.WebSocketComponent.Logic.Api;
 
 namespace backend.WebSocketComponent.Logic.Impl;
